Create build to fix display error
