+++ app/controllers/api/greetings_controller.rb
@@ -1,4 +1,4 @@
-class GreetingsController < ApplicationController
+class Api::GreetingsController < ApplicationController
   def random_greeting
     @greeting = Message.order('RANDOM()').first
     render json: { greeting: @greeting.greeting }
