--- app/javascript/components/App.js
@@ -1,20 +1,0 @@
-import React from "react";
-import { Provider } from "react-redux";
-// import ReactDOM from "react-dom";
-import { BrowserRouter, Routes, Route } from "react-router-dom";
-import store from "./redux/store";
-import Greeting from "./components/Greetings.js";
-
-function App() {
-  return (
-    <Provider store={store}>
-      <BrowserRouter>
-        <Routes>
-          <Route path="/" element={<Greeting />} />
-        </Routes>
-      </BrowserRouter>
-    </Provider>
-  );
-}
-
-export default App;
