+++ app/javascript/redux/reducers/greetingSlice.js
@@ -2,7 +2,7 @@ import { createSlice } from "@reduxjs/toolkit";
 import fetchGreeting from "../thunk";
 
 const initialState = {
-  greeting: "",
+  greeting: "Loading...",
   isLoading: true,
   error: false,
   errMsg: "",
@@ -14,11 +14,16 @@ const greetingSlice = createSlice({
   extraReducers: (builder) => {
     builder
       .addCase(fetchGreeting.pending, (state) => {
+        state.greeting = "";
         state.isLoading = true;
+        state.error = false;
+        state.errMsg = "";
       })
       .addCase(fetchGreeting.fulfilled, (state, action) => {
-        state.greeting = action.payload.message;
+        state.greeting = action.payload.greeting;
         state.isLoading = false;
+        state.error = false;
+        state.errMsg = "";
       })
       .addCase(fetchGreeting.rejected, (state, action) => {
         state.isLoading = false;
