+++ app/javascript/redux/thunk.js
@@ -1,6 +1,7 @@
 import { createAsyncThunk } from "@reduxjs/toolkit";
 import axios from "axios";
-const url = "/api/greetings/random_greeting";
+
+const url = "api/greetings/random_greeting";
 
 const fetchGreeting = createAsyncThunk(
   "greeting/fetchGreeting",
@@ -8,8 +9,8 @@ const fetchGreeting = createAsyncThunk(
     try {
       const response = await axios.get(url);
       return response.data;
-    } catch (e) {
-      return thunkAPI.rejectWithValue({ error: e.message });
+    } catch (error) {
+      return thunkAPI.rejectWithValue({ error: error.message });
     }
   }
 );
