--- app/views/greetings/random_greeting.html.erb
@@ -1,2 +1,0 @@
-<h1>Greetings#random_greeting</h1>
-<p>Find me in app/views/greetings/random_greeting.html.erb</p>
