+++ config/routes.rb
@@ -3,7 +3,7 @@ Rails.application.routes.draw do
   root 'root#index'
 
   namespace :api do
-  get 'greetings/random_greeting'
-  # get '/random_greeting', to: 'greetings#random_greeting'
-  end
+  get 'greetings/random_greeting', to: 'greetings#random_greeting'
+end
+
 end
