+++ db/seeds.rb
@@ -5,6 +5,11 @@
 #
 #   movies = Movie.create([{ name: "Star Wars" }, { name: "Lord of the Rings" }])
 #   Character.create(name: "Luke", movie: movies.first)
-5.times do
-  Message.create(greeting: "Hello, World!")
+# 5.times do
+#   Message.create(greeting: "Hello, World!")
+# end
+greeting_msgs = ['Hello World', 'Welcome to Rails React', 'Hello Webpack', 'Hello Redux', 'Hello World']
+
+greeting_msgs.each do |msg|
+  Message.create(greeting: msg)
 end
