+++ webpack.config.js
@@ -1,5 +1,5 @@
-const path    = require("path")
-const webpack = require("webpack")
+const path = require("path");
+const webpack = require("webpack");
 
 module.exports = {
   mode: "production",
@@ -18,7 +18,7 @@ module.exports = {
   },
   output: {
     filename: "[name].js",
-    sourceMapFilename: "[file].map",
+    sourceMapFilename: "[name].js.map",
     path: path.resolve(__dirname, "app/assets/builds"),
   },
   plugins: [
